nullexa/requirements.txt
python-socketio==5.8.0
flask==2.3.3
flask-socketio==5.3.4
eventlet>=0.40.0
python-engineio==4.5.1
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
blinker==1.6.2
greenlet>=3.0.3
bidict==0.22.1
python-dotenv==1.0.0
